create stock update sevlet
