+++ smd-distributors-automated-system/src/main/webapp/WEB-INF/views/stockM/stock_form.jsp
@@ -1,6 +1,6 @@
 <div class="col-4">
 		
-	<form action="../../StockAdding" method="Post">
+	<form action="../../StockUpdate" method="Post">
 				<!-- 	<div class="form-group">
 					<div class="dropdown show">
 						<a
@@ -27,11 +27,9 @@
 				<label for="exampleInputPassword1">Quntity</label> <input type="text"
 					class="form-control" id="exampleInputPassword1" name="quntity">
 			</div>
-			
-			
-			
-			
-			
-			<button type="Add Stock" class="btn btn-primary">Add stock</button>
+			<div>
+			<button name ="button" value="add" type="submit" name="id" class="btn btn-light">Add Stock</button>
+			<button name ="button" value="delete" type="submit" name="id" class="btn btn-light">Delete Stock</button>
+			</div>
 		</form>
 </div>
